refactor Sys (delete Sys2)
