+++ pages/statistics.tsx
@@ -183,8 +183,8 @@ const Home: NextPage = () => {
             <button onClick={setAllHandler}>All</button>
           </div>
           <div className={styles.orangeButton}>
-            <button onClick={showSalesHandler}>show Sales</button>
-            <button onClick={showFullSalesHandler}>show Statistic</button>
+            <button onClick={showSalesHandler}>Short report</button>
+            <button onClick={showFullSalesHandler}>Full statistic</button>
           </div>
           {resData === undefined || resData.length === 0 ? (
             <p>No data - empty result</p>
